requirements.txt
spotipy
typing
PyQt6
Flask
python-dotenv